+++ source/packages/@aws-accelerator/accelerator/lib/stacks/applications-stack.ts
@@ -532,6 +532,7 @@ export class ApplicationsStack extends AcceleratorStack {
         lambdaKey,
         cloudWatchLogKmsKey,
         cloudWatchLogRetentionInDays,
+        maxInstanceLifetime: appConfigItem.autoscaling.maxInstanceLifetime ?? undefined,
       },
     );