+++ source/packages/@aws-accelerator/accelerator/lib/stacks/network-stacks/network-associations-gwlb-stack/network-associations-gwlb-stack.ts
@@ -707,6 +707,7 @@ export class NetworkAssociationsGwlbStack extends NetworkStack {
       healthCheckType: group.healthCheckType,
       targetGroups: group.targetGroups ? this.processTargetGroups(group.targetGroups) : undefined,
       subnets: this.processSubnets(group.subnets, vpc),
+      maxInstanceLifetime: group.maxInstanceLifetime,
     };
   }