+++ source/packages/@aws-accelerator/constructs/lib/aws-ec2/firewall-asg.ts
@@ -66,6 +66,7 @@ export class FirewallAutoScalingGroup extends Firewall implements IFirewallAutoS
       lambdaKey: props.lambdaKey,
       cloudWatchLogKmsKey: props.cloudWatchLogKmsKey,
       cloudWatchLogRetentionInDays: props.cloudWatchLogRetentionInDays,
+      maxInstanceLifetime: props.autoscaling.maxInstanceLifetime,
       nagSuppressionPrefix: `${id}/Resource`,
     });